+++ streaming/song.py
@@ -13,21 +13,20 @@ class Song:
         basename (str): The filename of the song file.
     """
 
-    def __init__(self, file: str,  song_name = "", song_artist: str = "", song_requested_by: str = ""):
+    def __init__(self, file: str, song_name: str = "", song_artist: str = "", song_requested_by: str = ""):
         self.is_playing = False
         self.is_paused = False
         self.is_stopped = True
         self.requested_by = song_requested_by
         self.file = file
         self.basename = os.path.basename(self.file)
-        self.song_name =  song_name
+        self.song_name = song_name
         self.artist = song_artist
 
         if len(song_name) == 0:
             name: str = self.basename.split("/")[-1].split(".")
             name = (".".join(name[:len(name) - 1]).replace("_", " ").replace("-", " - "))
             self.set_song_name(name)
-
 
     def is_request(self):
         """